refactor: move ensurePasswordValidity to AuthenticationValidator
